refactor: type navigators in the application
