+++ src/navigation/stack/src/HomeStack.tsx
@@ -4,9 +4,10 @@ import {createNativeStackNavigator} from "@react-navigation/native-stack";
 
 import {HomeAppBar, NotificationsAppBar} from "@/components";
 import {stackScreenOptions} from "@/navigation/config";
+import {HomeTabParamList} from "@/navigation/navigation";
 import {HomeScreen, NotificationsScreen} from "@/screens/home";
 
-const Stack = createNativeStackNavigator();
+const Stack = createNativeStackNavigator<HomeTabParamList>();
 
 export default function HomeStack() {
   return (
