+++ src/navigation/stack/src/MyPageStack.tsx
@@ -4,9 +4,10 @@ import {createNativeStackNavigator} from "@react-navigation/native-stack";
 
 import {MyPageAppBar} from "@/components";
 import {stackScreenOptions} from "@/navigation/config";
+import {MyPageTabParamList} from "@/navigation/navigation";
 import {MyPageMainScreen} from "@/screens/mypage";
 
-const Stack = createNativeStackNavigator();
+const Stack = createNativeStackNavigator<MyPageTabParamList>();
 
 export default function MyPageStack() {
   return (
