+++ src/navigation/stack/src/WriteStack.tsx
@@ -3,15 +3,16 @@ import * as React from "react";
 import {createNativeStackNavigator} from "@react-navigation/native-stack";
 
 import {stackScreenOptions} from "@/navigation/config";
+import {WriteTabParamList} from "@/navigation/navigation";
 import WriteScreen from "@/screens/write/Main";
 
-const Stack = createNativeStackNavigator();
+const Stack = createNativeStackNavigator<WriteTabParamList>();
 
 export default function WriteStack() {
   return (
-    <Stack.Navigator initialRouteName="Write" screenOptions={stackScreenOptions}>
+    <Stack.Navigator initialRouteName="WriteScreen" screenOptions={stackScreenOptions}>
       <Stack.Group>
-        <Stack.Screen name="Write" component={WriteScreen} />
+        <Stack.Screen name="WriteScreen" component={WriteScreen} />
       </Stack.Group>
     </Stack.Navigator>
   );
